+++ packages/reactackle-core/package.json
@@ -24,8 +24,7 @@
     "lodash.merge": "^4.6.0",
     "lodash.omit": "^4.5.0",
     "lodash.pick": "^4.4.0",
-    "prop-types": "^15.5.10",
-    "react-test-renderer": "^15.6.1"
+    "prop-types": "^15.5.10"
   },
   "peerDependencies": {
     "react": "^15.6.1",
