+++ packages/reactackle-core/src/utils.js
@@ -1,6 +1,5 @@
 export * from './utils/misc';
 export * from './utils/polyfills';
-export * from './utils/jest';
 export * from './utils/index';
 export * from './utils/create-broadcast';
 export * from './utils/contextChannel';
