+++ packages/reactackle-input-autocomplete/package.json
@@ -31,7 +31,8 @@
     "babel-cli": "^6.24.1",
     "react-test-renderer": "^15.6.1",
     "reactackle-core": "^0.0.3-beta.1",
-    "rimraf": "^2.6.1"
+    "rimraf": "^2.6.1",
+    "reactackle-test-utils": "^0.0.3-beta.1"
   },
   "repository": {
     "type": "git",
