+++ packages/reactackle-selectbox/__tests__/select-box.test.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import renderer from 'react-test-renderer';
-import { findByType } from 'reactackle-core';
+import { findByType } from 'reactackle-test-utils';
 
 import {
   OptionCustom,
