+++ packages/reactackle-test-utils/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "reactackle-checkbox",
+  "name": "reactackle-test-utils",
   "version": "0.0.3-beta.1",
-  "description": "Checkbox component of the Reactackle ui library",
+  "description": "Test utils of the Reactackle ui library",
   "scripts": {
     "prepublishOnly": "rimraf ./dist && npx babel ./src --out-dir ./dist"
   },
@@ -17,22 +17,14 @@
     "dist"
   ],
   "dependencies": {
-    "prop-types": "^15.5.10",
-    "reactackle-icon": "^0.0.3-beta.1",
-    "reactackle-tooltip": "^0.0.3-beta.1",
-    "reactackle-tooltip-icon": "^0.0.3-beta.1"
+    "react-test-renderer": "^15.6.1"
   },
   "peerDependencies": {
-    "react": "^15.6.1",
-    "reactackle-core": "^0.0.3-beta.1",
-    "styled-components": "^2.1.1"
+    "react": "^15.6.1"
   },
   "devDependencies": {
     "babel-cli": "^6.24.1",
-    "react-test-renderer": "^15.6.1",
-    "reactackle-core": "^0.0.3-beta.1",
-    "rimraf": "^2.6.1",
-    "reactackle-test-utils": "^0.0.3-beta.1"
+    "rimraf": "^2.6.1"
   },
   "repository": {
     "type": "git",
