+++ packages/reactackle-text-field/package.json
@@ -34,7 +34,8 @@
     "enzyme-to-json": "^1.5.1",
     "react-test-renderer": "^15.6.1",
     "reactackle-core": "^0.0.3-beta.1",
-    "rimraf": "^2.6.1"
+    "rimraf": "^2.6.1",
+    "reactackle-test-utils": "^0.0.3-beta.1"
   },
   "repository": {
     "type": "git",
